fixing bugs I didn't know I had

+++ src/main/java/magicjinn/theblockkeepsticking/mixin/CampfireBlockEntityMixin.java
@@ -28,7 +28,7 @@ public class CampfireBlockEntityMixin implements TickingBlockAccessor {
         for (int i = 0; i < itemsBeingCooked.size(); i++) {
             ItemStack item = itemsBeingCooked.get(i);
             if (!item.isEmpty()) {
-                cookingTimes[i] += (long) ticksToSimulate;
+                cookingTimes[i] += (int) ticksToSimulate;
                 changed = true;
             }
         }
